add EventBackendModule
This module is an helper for building backend module without dca.

A GET 'action' param can be use to select the method get_THE_VALUE.
A POST action array param can be used to, letting you launching an
action by the name of a button, with several button in the same form.

eg:
...
<input type="submit" name="action[delete]" value="delete" />
<input type="submit" name="action[update]" value="update" />
...

Default action is index(). Any other action redirect to index after
processing if html format is requested ( default ), or die in a json
response if json format is requested.