add caching for getters/setters
In EModel, EventBackendModule and RoutedModule, the result of a getter
is cached the first time it is called.

You can prevent this by adding the key name to the uncachable array.

eg:

$this->uncachable[] = 'name';
protected function getName(){ ... }

If the setter is not defined, the default is to simply pass the value to
the cache array.

$this->name = 'foo',
=> $this->arrCache[ 'name' ] is now 'foo'.

So, why to use this instead of the default Typolight __get and __set
$arrValue array?

Because you can always define an getSomething or setSomething method to do
more complex processing. Typolight default is to simply assigned what you
pass to the array.

You can still use default getter and setter if you don't need complex
processing. Simply not declare getSomething nor setSomething for the
something attribute.