Let HardRoutesList load routes in database.
In the backend module that lists hard routes ( routes defined in module
config files ), we can now choose to load a given route in the database.

This will assure that modules using framework are distribuable. If
someone use such a module and want to modify routes without knowledge of
how modules work, he may just load the routes in database and edit them
in backend.

Btw, HardRoutesList.php now use EventBackendModule and can be a good
example for it usage.