add WidgetEFileTree
eFileTree works as fileTree, but doesn't blow off your path argument if
the current user has a filemount. It will try to respect it, given that
the path is accessible in the user's filemount.

Additionally, a link is displayed in the widget to upload a new file at
path, if path is given.