clause in associations in EModel, cleanup
A where and limit clause can be used when accessing an association, as
in first(), last() and getAll().

A cleanup was also added for relation tables. When you call delete() on
a EModel, all its entries in the relation tables are deleted.

To use this in dca, add an ondelete_callback :
array( '[TheClass]', 'cleanupAssociation' )