enhance RoutedModule#paginate()
You can now pass an EModel instance in place of the collection for
paginate(), with optional sql clauses. The pagination will be realized
by calling getAll() with clauses and the shortest limit on this object,
in order to enhance perfs.

The pagination links generation has been extracted in
preparePagination(), so you can use your own filter method and simply
pass in the total of items, and the current page index.