let routes be more fail complient
When a the first argument of Route::compose() doesn't match any
route, it is checked against page aliases.

If a page alias match, a link to this page with params in standard
shape is returned.

If it doesn't match nor route nor alias, a link to the current page
is sent with given params.