ADD ModuleRoutedBreadcrumb
Let you add the current action at the end of the breadcrumb, and the
link to the index action right before it. This way, virtual pages
generated by FrontendController are not shown as the same page.

In order for this to work, you have to specify a controller name in
the static params of the current route, and set the
$GLOBALS[ 'TL_LANG' ][ 'TL_MSC' ][ 'routes' ][ <controller> ][ <action> ]
value.